requirements.txt
discord.py
tomlkit